Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -3,8 +3,8 @@
 
 
 int main() {
-   float a,b,c;
-   scanf("%lf %lf %lf",&a,&b,&c);
+   int a,b,c;
+   scanf("%i %i %i",&a,&b,&c);
    float d = (a+b+c)/3;
    printf("Average: %.2lf",d);
